+++ sample/DAL/StudentInsert.cs
@@ -4,17 +4,19 @@ using System.Linq;
 using System.Web;
 using System.Web.ModelBinding;
 using sample.Models;
+using sample.DAL;
 
 namespace sample.DAL
 {
     public class StudentInsert
     {
-        private  SchoolContext db = new SchoolContext();
+       
         public void InsertData(StudentModels studentModels)
         {
+            SchoolContext context = new SchoolContext();
             studentModels.EnrollmentDate = Convert.ToDateTime(DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss"));
-            db.StudentModelses.Add(studentModels);
-            db.SaveChanges();
+            context.StudentModels.Add(studentModels);
+            context.SaveChanges();
         }
 
     }
