+++ sample/Views/Student/Index.cshtml
@@ -13,18 +13,18 @@ table thead tr th {
 <table id="Student_List" style="border: 1px solid black; width: 500px;  text-align: center">
     <tr>
         <td>StudentID</td>
-        <td><input id="StudentModelsID" type="number" maxlength="10"/></td>
+        <td><input id="studentModelsID" type="number" maxlength="10"/></td>
     </tr>
     <tr>
         <td>LastName</td>
-        <td><input id="LastName"/></td>
+        <td><input id="lastName"/></td>
     </tr>
     <tr>
         <td>FirstMidName</td>
-        <td><input id="FirstMidName"/></td>
+        <td><input id="firstMidName"/></td>
     </tr>
 </table>
-<p><input type="button" value="저장" id="InsertDB"/></p>
+<p><input type="button" value="저장" id="insertDB"/></p>
 <br/>
 
 <table style="border: 1px solid black; width: 800px; border-spacing: 20px; text-align: center">
